+++ src/BarChart/BarSeries/BarLabel.tsx
@@ -1,7 +1,7 @@
 import React, { FC, PureComponent, useMemo } from 'react';
 import { ChartInternalShallowDataShape, Direction } from '../../common/data';
 import { BarType } from './Bar';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { DEFAULT_TRANSITION } from '../../common/Motion';
 
 export interface BarLabelProps {