+++ src/Heatmap/HeatmapSeries/HeatmapCell.tsx
@@ -15,7 +15,7 @@ import {
 } from '../../common/utils/functions';
 import chroma from 'chroma-js';
 import classNames from 'classnames';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { DEFAULT_TRANSITION } from '../../common/Motion';
 import { ChartInternalShallowDataShape } from '../../common/data';
 import css from './HeatmapCell.module.css';