+++ src/PieChart/PieArcSeries/PieArc.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement, useState, FC, useRef, useMemo } from 'react';
 import chroma from 'chroma-js';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { CloneElement } from 'rdk';
 import { ArcData } from '../PieChart';
 import { ChartTooltip, ChartTooltipProps } from '../../common/Tooltip';