+++ src/RadialScatterPlot/RadialScatterSeries/RadialScatterPoint.tsx
@@ -11,7 +11,7 @@ import { radialLine } from 'd3-shape';
 import classNames from 'classnames';
 import { ChartTooltip, ChartTooltipProps } from '../../common/Tooltip';
 import { CloneElement } from 'rdk';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { DEFAULT_TRANSITION } from '../../common/Motion';
 import { schemes } from '../../common/color';
 import css from './RadialScatterPoint.module.css';