+++ src/ScatterPlot/ScatterSeries/ScatterPoint.tsx
@@ -15,7 +15,7 @@ import {
   constructFunctionProps,
   PropFunctionTypes
 } from '../../common/utils/functions';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { DEFAULT_TRANSITION } from '../../common/Motion';
 import { schemes, getColor, ColorSchemeType } from '../../common/color';
 import css from './ScatterPoint.module.css';