+++ src/VennDiagram/VennLabel.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { wrapText } from '../common/utils/wrapText';
 import { DEFAULT_TRANSITION } from '../common/Motion';