+++ src/VennDiagram/VennSeries.tsx
@@ -10,7 +10,7 @@ import { IVennLayout } from '@upsetjs/venn.js';
 import { ColorSchemeType, getColor } from '../common/color';
 import { VennArc, VennArcProps } from './VennArc';
 import { VennLabel, VennLabelProps } from './VennLabel';
-import { motion } from 'framer-motion/dist/framer-motion';
+import { motion } from 'framer-motion';
 import { CloneElement } from 'rdk';
 import chroma from 'chroma-js';
 import { VennOuterLabel, VennOuterLabelProps } from './VennOuterLabel';